Struct aws_sdk_quicksight::client::fluent_builders::RegisterUser

pub struct RegisterUser<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> RegisterUser<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<RegisterUserOutput, SdkError<RegisterUserError>> where
R::Policy: SmithyRetryPolicy<RegisterUserInputOperationOutputAlias, RegisterUserOutput, RegisterUserError, RegisterUserInputOperationRetryAlias>,

pub fn identity_type(self, inp: IdentityType) -> Self

Amazon QuickSight supports several ways of managing the identity of users. This parameter accepts two values:

IAM: A user whose identity maps to an existing IAMuser or role.

QUICKSIGHT: A user whose identity is owned and managed internally by Amazon QuickSight.

pub fn set_identity_type(self, input: Option<IdentityType>) -> Self

pub fn email(self, inp: impl Into<String>) -> Self

The email address of the user that you want to register.

pub fn set_email(self, input: Option<String>) -> Self

pub fn user_role(self, inp: UserRole) -> Self

The Amazon QuickSight role for the user. The user role can be one of the following:

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon QuickSight settings.

RESTRICTED_READER: This role isn't currently available for use.

RESTRICTED_AUTHOR: This role isn't currently available for use.

pub fn set_user_role(self, input: Option<UserRole>) -> Self

pub fn iam_arn(self, inp: impl Into<String>) -> Self

The ARN of the IAMuser or role that you are registering with Amazon QuickSight.

pub fn set_iam_arn(self, input: Option<String>) -> Self

pub fn session_name(self, inp: impl Into<String>) -> Self

You need to use this parameter only when you register one or more users using an assumed IAMrole. You don't need to provide the session name for other scenarios, for example when you are registering an IAMuser or an Amazon QuickSight user. You can register multiple users using the same IAMrole if each user has a different session name. For more information on assuming IAMroles, see assume-role in the AWS CLI Reference.

pub fn set_session_name(self, input: Option<String>) -> Self

pub fn aws_account_id(self, inp: impl Into<String>) -> Self

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

pub fn set_aws_account_id(self, input: Option<String>) -> Self

pub fn namespace(self, inp: impl Into<String>) -> Self

The namespace. Currently, you should set this to default.

pub fn set_namespace(self, input: Option<String>) -> Self

pub fn user_name(self, inp: impl Into<String>) -> Self

The Amazon QuickSight user name that you want to create for the user you are registering.

pub fn set_user_name(self, input: Option<String>) -> Self

pub fn custom_permissions_name(self, inp: impl Into<String>) -> Self

(Enterprise edition only) The name of the custom permissions profile that you want to assign to this user. Customized permissions allows you to control a user's access by restricting access the following operations:

Create and update data sources

Create and update datasets

Create and update email reports

Subscribe to email reports

To add custom permissions to an existing user, use UpdateUser instead.

A set of custom permissions includes any combination of these restrictions. Currently, you need to create the profile names for custom permission sets by using the Amazon QuickSight console. Then, you use the RegisterUser API operation to assign the named set of permissions to a Amazon QuickSight user.

Amazon QuickSight custom permissions are applied through IAMpolicies. Therefore, they override the permissions typically granted by assigning Amazon QuickSight users to one of the default security cohorts in Amazon QuickSight (admin, author, reader).

This feature is available only to Amazon QuickSight Enterprise edition subscriptions.

pub fn set_custom_permissions_name(self, input: Option<String>) -> Self

pub fn external_login_federation_provider_type(
self,
inp: impl Into<String>
) -> Self

The type of supported external login provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. The type of supported external login provider can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com. When choosing the COGNITO provider type, don’t use the "CustomFederationProviderUrl" parameter which is only needed when the external provider is custom.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider. When choosing CUSTOM_OIDC type, use the CustomFederationProviderUrl parameter to provide the custom OIDC provider URL.

pub fn set_external_login_federation_provider_type(
self,
input: Option<String>
) -> Self

pub fn custom_federation_provider_url(self, inp: impl Into<String>) -> Self

The URL of the custom OpenID Connect (OIDC) provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. This parameter should only be used when ExternalLoginFederationProviderType parameter is set to CUSTOM_OIDC.

pub fn set_custom_federation_provider_url(self, input: Option<String>) -> Self

pub fn external_login_id(self, inp: impl Into<String>) -> Self

The identity ID for a user in the external login provider.

pub fn set_external_login_id(self, input: Option<String>) -> Self