Struct aws_sdk_quicksight::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for QuickSight_20180401.

This client allows ergonomic access to a QuickSight_20180401-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn cancel_ingestion(&self) -> CancelIngestion<C, M, R>

pub fn create_account_customization(
&self
) -> CreateAccountCustomization<C, M, R>

pub fn create_analysis(&self) -> CreateAnalysis<C, M, R>

pub fn create_dashboard(&self) -> CreateDashboard<C, M, R>

pub fn create_data_set(&self) -> CreateDataSet<C, M, R>

pub fn create_data_source(&self) -> CreateDataSource<C, M, R>

pub fn create_folder(&self) -> CreateFolder<C, M, R>

pub fn create_folder_membership(&self) -> CreateFolderMembership<C, M, R>

pub fn create_group(&self) -> CreateGroup<C, M, R>

pub fn create_group_membership(&self) -> CreateGroupMembership<C, M, R>

pub fn create_iam_policy_assignment(&self) -> CreateIAMPolicyAssignment<C, M, R>

pub fn create_ingestion(&self) -> CreateIngestion<C, M, R>

pub fn create_namespace(&self) -> CreateNamespace<C, M, R>

pub fn create_template(&self) -> CreateTemplate<C, M, R>

pub fn create_template_alias(&self) -> CreateTemplateAlias<C, M, R>

pub fn create_theme(&self) -> CreateTheme<C, M, R>

pub fn create_theme_alias(&self) -> CreateThemeAlias<C, M, R>

pub fn delete_account_customization(
&self
) -> DeleteAccountCustomization<C, M, R>

pub fn delete_analysis(&self) -> DeleteAnalysis<C, M, R>

pub fn delete_dashboard(&self) -> DeleteDashboard<C, M, R>

pub fn delete_data_set(&self) -> DeleteDataSet<C, M, R>

pub fn delete_data_source(&self) -> DeleteDataSource<C, M, R>

pub fn delete_folder(&self) -> DeleteFolder<C, M, R>

pub fn delete_folder_membership(&self) -> DeleteFolderMembership<C, M, R>

pub fn delete_group(&self) -> DeleteGroup<C, M, R>

pub fn delete_group_membership(&self) -> DeleteGroupMembership<C, M, R>

pub fn delete_iam_policy_assignment(&self) -> DeleteIAMPolicyAssignment<C, M, R>

pub fn delete_namespace(&self) -> DeleteNamespace<C, M, R>

pub fn delete_template(&self) -> DeleteTemplate<C, M, R>

pub fn delete_template_alias(&self) -> DeleteTemplateAlias<C, M, R>

pub fn delete_theme(&self) -> DeleteTheme<C, M, R>

pub fn delete_theme_alias(&self) -> DeleteThemeAlias<C, M, R>

pub fn delete_user(&self) -> DeleteUser<C, M, R>

pub fn delete_user_by_principal_id(&self) -> DeleteUserByPrincipalId<C, M, R>

pub fn describe_account_customization(
&self
) -> DescribeAccountCustomization<C, M, R>

pub fn describe_account_settings(&self) -> DescribeAccountSettings<C, M, R>

pub fn describe_analysis(&self) -> DescribeAnalysis<C, M, R>

pub fn describe_analysis_permissions(
&self
) -> DescribeAnalysisPermissions<C, M, R>

pub fn describe_dashboard(&self) -> DescribeDashboard<C, M, R>

pub fn describe_dashboard_permissions(
&self
) -> DescribeDashboardPermissions<C, M, R>

pub fn describe_data_set(&self) -> DescribeDataSet<C, M, R>

pub fn describe_data_set_permissions(
&self
) -> DescribeDataSetPermissions<C, M, R>

pub fn describe_data_source(&self) -> DescribeDataSource<C, M, R>

pub fn describe_data_source_permissions(
&self
) -> DescribeDataSourcePermissions<C, M, R>

pub fn describe_folder(&self) -> DescribeFolder<C, M, R>

pub fn describe_folder_permissions(&self) -> DescribeFolderPermissions<C, M, R>

pub fn describe_folder_resolved_permissions(
&self
) -> DescribeFolderResolvedPermissions<C, M, R>

pub fn describe_group(&self) -> DescribeGroup<C, M, R>

pub fn describe_iam_policy_assignment(
&self
) -> DescribeIAMPolicyAssignment<C, M, R>

pub fn describe_ingestion(&self) -> DescribeIngestion<C, M, R>

pub fn describe_namespace(&self) -> DescribeNamespace<C, M, R>

pub fn describe_template(&self) -> DescribeTemplate<C, M, R>

pub fn describe_template_alias(&self) -> DescribeTemplateAlias<C, M, R>

pub fn describe_template_permissions(
&self
) -> DescribeTemplatePermissions<C, M, R>

pub fn describe_theme(&self) -> DescribeTheme<C, M, R>

pub fn describe_theme_alias(&self) -> DescribeThemeAlias<C, M, R>

pub fn describe_theme_permissions(&self) -> DescribeThemePermissions<C, M, R>

pub fn describe_user(&self) -> DescribeUser<C, M, R>

pub fn generate_embed_url_for_anonymous_user(
&self
) -> GenerateEmbedUrlForAnonymousUser<C, M, R>

pub fn generate_embed_url_for_registered_user(
&self
) -> GenerateEmbedUrlForRegisteredUser<C, M, R>

pub fn get_dashboard_embed_url(&self) -> GetDashboardEmbedUrl<C, M, R>

pub fn get_session_embed_url(&self) -> GetSessionEmbedUrl<C, M, R>

pub fn list_analyses(&self) -> ListAnalyses<C, M, R>

pub fn list_dashboards(&self) -> ListDashboards<C, M, R>

pub fn list_dashboard_versions(&self) -> ListDashboardVersions<C, M, R>

pub fn list_data_sets(&self) -> ListDataSets<C, M, R>

pub fn list_data_sources(&self) -> ListDataSources<C, M, R>

pub fn list_folder_members(&self) -> ListFolderMembers<C, M, R>

pub fn list_folders(&self) -> ListFolders<C, M, R>

pub fn list_group_memberships(&self) -> ListGroupMemberships<C, M, R>

pub fn list_groups(&self) -> ListGroups<C, M, R>

pub fn list_iam_policy_assignments(&self) -> ListIAMPolicyAssignments<C, M, R>

pub fn list_iam_policy_assignments_for_user(
&self
) -> ListIAMPolicyAssignmentsForUser<C, M, R>

pub fn list_ingestions(&self) -> ListIngestions<C, M, R>

pub fn list_namespaces(&self) -> ListNamespaces<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_template_aliases(&self) -> ListTemplateAliases<C, M, R>

pub fn list_templates(&self) -> ListTemplates<C, M, R>

pub fn list_template_versions(&self) -> ListTemplateVersions<C, M, R>

pub fn list_theme_aliases(&self) -> ListThemeAliases<C, M, R>

pub fn list_themes(&self) -> ListThemes<C, M, R>

pub fn list_theme_versions(&self) -> ListThemeVersions<C, M, R>

pub fn list_user_groups(&self) -> ListUserGroups<C, M, R>

pub fn list_users(&self) -> ListUsers<C, M, R>

pub fn register_user(&self) -> RegisterUser<C, M, R>

pub fn restore_analysis(&self) -> RestoreAnalysis<C, M, R>

pub fn search_analyses(&self) -> SearchAnalyses<C, M, R>

pub fn search_dashboards(&self) -> SearchDashboards<C, M, R>

pub fn search_folders(&self) -> SearchFolders<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_account_customization(
&self
) -> UpdateAccountCustomization<C, M, R>

pub fn update_account_settings(&self) -> UpdateAccountSettings<C, M, R>

pub fn update_analysis(&self) -> UpdateAnalysis<C, M, R>

pub fn update_analysis_permissions(&self) -> UpdateAnalysisPermissions<C, M, R>

pub fn update_dashboard(&self) -> UpdateDashboard<C, M, R>

pub fn update_dashboard_permissions(
&self
) -> UpdateDashboardPermissions<C, M, R>

pub fn update_dashboard_published_version(
&self
) -> UpdateDashboardPublishedVersion<C, M, R>

pub fn update_data_set(&self) -> UpdateDataSet<C, M, R>

pub fn update_data_set_permissions(&self) -> UpdateDataSetPermissions<C, M, R>

pub fn update_data_source(&self) -> UpdateDataSource<C, M, R>

pub fn update_data_source_permissions(
&self
) -> UpdateDataSourcePermissions<C, M, R>

pub fn update_folder(&self) -> UpdateFolder<C, M, R>

pub fn update_folder_permissions(&self) -> UpdateFolderPermissions<C, M, R>

pub fn update_group(&self) -> UpdateGroup<C, M, R>

pub fn update_iam_policy_assignment(&self) -> UpdateIAMPolicyAssignment<C, M, R>

pub fn update_template(&self) -> UpdateTemplate<C, M, R>

pub fn update_template_alias(&self) -> UpdateTemplateAlias<C, M, R>

pub fn update_template_permissions(&self) -> UpdateTemplatePermissions<C, M, R>

pub fn update_theme(&self) -> UpdateTheme<C, M, R>

pub fn update_theme_alias(&self) -> UpdateThemeAlias<C, M, R>

pub fn update_theme_permissions(&self) -> UpdateThemePermissions<C, M, R>

pub fn update_user(&self) -> UpdateUser<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self