Enum aws_sdk_quicksight::Error

#[non_exhaustive]
pub enum Error {

    AccessDeniedException(AccessDeniedException),
    ConcurrentUpdatingException(ConcurrentUpdatingException),
    ConflictException(ConflictException),
    DomainNotWhitelistedException(DomainNotWhitelistedException),
    IdentityTypeNotSupportedException(IdentityTypeNotSupportedException),
    InternalFailureException(InternalFailureException),
    InvalidNextTokenException(InvalidNextTokenException),
    InvalidParameterValueException(InvalidParameterValueException),
    LimitExceededException(LimitExceededException),
    PreconditionNotMetException(PreconditionNotMetException),
    QuickSightUserNotFoundException(QuickSightUserNotFoundException),
    ResourceExistsException(ResourceExistsException),
    ResourceNotFoundException(ResourceNotFoundException),
    ResourceUnavailableException(ResourceUnavailableException),
    SessionLifetimeInMinutesInvalidException(SessionLifetimeInMinutesInvalidException),
    ThrottlingException(ThrottlingException),
    UnsupportedPricingPlanException(UnsupportedPricingPlanException),
    UnsupportedUserEditionException(UnsupportedUserEditionException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),

}

All possible error types for this service.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

You don't have access to this item. The provided credentials couldn't be validated. You might not be authorized to carry out the request. Make sure that your account is authorized to use the Amazon QuickSight service, that your policies have the correct permissions, and that you are using the correct access keys.

Tuple Fields of AccessDeniedException

0: AccessDeniedException

ConcurrentUpdatingException(ConcurrentUpdatingException)

A resource is already in a state that indicates an operation is happening that must complete before a new update can be applied.

Tuple Fields of ConcurrentUpdatingException

0: ConcurrentUpdatingException

ConflictException(ConflictException)

Updating or deleting a resource can cause an inconsistent state.

Tuple Fields of ConflictException

0: ConflictException

DomainNotWhitelistedException(DomainNotWhitelistedException)

The domain specified isn't on the allow list. All domains for embedded dashboards must be added to the approved list by an Amazon QuickSight admin.

Tuple Fields of DomainNotWhitelistedException

0: DomainNotWhitelistedException

IdentityTypeNotSupportedException(IdentityTypeNotSupportedException)

The identity type specified isn't supported. Supported identity types include IAM and QUICKSIGHT.

Tuple Fields of IdentityTypeNotSupportedException

0: IdentityTypeNotSupportedException

InternalFailureException(InternalFailureException)

An internal failure occurred.

Tuple Fields of InternalFailureException

0: InternalFailureException

InvalidNextTokenException(InvalidNextTokenException)

The NextToken value isn't valid.

Tuple Fields of InvalidNextTokenException

0: InvalidNextTokenException

InvalidParameterValueException(InvalidParameterValueException)

One or more parameters has a value that isn't valid.

Tuple Fields of InvalidParameterValueException

0: InvalidParameterValueException

LimitExceededException(LimitExceededException)

A limit is exceeded.

Tuple Fields of LimitExceededException

0: LimitExceededException

PreconditionNotMetException(PreconditionNotMetException)

One or more preconditions aren't met.

Tuple Fields of PreconditionNotMetException

0: PreconditionNotMetException

QuickSightUserNotFoundException(QuickSightUserNotFoundException)

The user with the provided name isn't found. This error can happen in any operation that requires finding a user based on a provided user name, such as DeleteUser, DescribeUser, and so on.

Tuple Fields of QuickSightUserNotFoundException

0: QuickSightUserNotFoundException

ResourceExistsException(ResourceExistsException)

The resource specified already exists.

Tuple Fields of ResourceExistsException

0: ResourceExistsException

ResourceNotFoundException(ResourceNotFoundException)

One or more resources can't be found.

Tuple Fields of ResourceNotFoundException

0: ResourceNotFoundException

ResourceUnavailableException(ResourceUnavailableException)

This resource is currently unavailable.

Tuple Fields of ResourceUnavailableException

0: ResourceUnavailableException

SessionLifetimeInMinutesInvalidException(SessionLifetimeInMinutesInvalidException)

The number of minutes specified for the lifetime of a session isn't valid. The session lifetime must be 15-600 minutes.

Tuple Fields of SessionLifetimeInMinutesInvalidException

0: SessionLifetimeInMinutesInvalidException

ThrottlingException(ThrottlingException)

Access is throttled.

Tuple Fields of ThrottlingException

0: ThrottlingException

UnsupportedPricingPlanException(UnsupportedPricingPlanException)

This error indicates that you are calling an embedding operation in Amazon QuickSight without the required pricing plan on your Amazon Web Services account. Before you can use embedding for anonymous users, a Amazon QuickSight administrator needs to add capacity pricing to Amazon QuickSight. You can do this on the Manage Amazon QuickSight page.

After capacity pricing is added, you can use the GetDashboardEmbedUrl API operation with the --identity-type ANONYMOUS option.

Tuple Fields of UnsupportedPricingPlanException

0: UnsupportedPricingPlanException

UnsupportedUserEditionException(UnsupportedUserEditionException)

This error indicates that you are calling an operation on an Amazon QuickSight subscription where the edition doesn't include support for that operation. Amazon Amazon QuickSight currently has Standard Edition and Enterprise Edition. Not every operation and capability is available in every edition.

Tuple Fields of UnsupportedUserEditionException

0: UnsupportedUserEditionException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unhandled error occurred.

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>