Module aws_sdk_quicksight::error

Modules

access_denied_exception

See AccessDeniedException

concurrent_updating_exception

See ConcurrentUpdatingException

conflict_exception

See ConflictException

domain_not_whitelisted_exception

See DomainNotWhitelistedException

identity_type_not_supported_exception

See IdentityTypeNotSupportedException

internal_failure_exception

See InternalFailureException

invalid_next_token_exception

See InvalidNextTokenException

invalid_parameter_value_exception

See InvalidParameterValueException

limit_exceeded_exception

See LimitExceededException

precondition_not_met_exception

See PreconditionNotMetException

quick_sight_user_not_found_exception

See QuickSightUserNotFoundException

resource_exists_exception

See ResourceExistsException

resource_not_found_exception

See ResourceNotFoundException

resource_unavailable_exception

See ResourceUnavailableException

session_lifetime_in_minutes_invalid_exception

See SessionLifetimeInMinutesInvalidException

throttling_exception

See ThrottlingException

unsupported_pricing_plan_exception

See UnsupportedPricingPlanException

unsupported_user_edition_exception

See UnsupportedUserEditionException

Structs

AccessDeniedException

You don't have access to this item. The provided credentials couldn't be

CancelIngestionError

ConcurrentUpdatingException

A resource is already in a state that indicates an operation is happening that must complete

ConflictException

Updating or deleting a resource can cause an inconsistent state.

CreateAccountCustomizationError

CreateAnalysisError

CreateDashboardError

CreateDataSetError

CreateDataSourceError

CreateFolderError

CreateFolderMembershipError

CreateGroupError

CreateGroupMembershipError

CreateIAMPolicyAssignmentError

CreateIngestionError

CreateNamespaceError

CreateTemplateAliasError

CreateTemplateError

CreateThemeAliasError

CreateThemeError

DeleteAccountCustomizationError

DeleteAnalysisError

DeleteDashboardError

DeleteDataSetError

DeleteDataSourceError

DeleteFolderError

DeleteFolderMembershipError

DeleteGroupError

DeleteGroupMembershipError

DeleteIAMPolicyAssignmentError

DeleteNamespaceError

DeleteTemplateAliasError

DeleteTemplateError

DeleteThemeAliasError

DeleteThemeError

DeleteUserByPrincipalIdError

DeleteUserError

DescribeAccountCustomizationError

DescribeAccountSettingsError

DescribeAnalysisError

DescribeAnalysisPermissionsError

DescribeDashboardError

DescribeDashboardPermissionsError

DescribeDataSetError

DescribeDataSetPermissionsError

DescribeDataSourceError

DescribeDataSourcePermissionsError

DescribeFolderError

DescribeFolderPermissionsError

DescribeFolderResolvedPermissionsError

DescribeGroupError

DescribeIAMPolicyAssignmentError

DescribeIngestionError

DescribeNamespaceError

DescribeTemplateAliasError

DescribeTemplateError

DescribeTemplatePermissionsError

DescribeThemeAliasError

DescribeThemeError

DescribeThemePermissionsError

DescribeUserError

DomainNotWhitelistedException

The domain specified isn't on the allow list. All domains for embedded dashboards must be

GenerateEmbedUrlForAnonymousUserError

GenerateEmbedUrlForRegisteredUserError

GetDashboardEmbedUrlError

GetSessionEmbedUrlError

IdentityTypeNotSupportedException

The identity type specified isn't supported. Supported identity types include

InternalFailureException

An internal failure occurred.

InvalidNextTokenException

The NextToken value isn't valid.

InvalidParameterValueException

One or more parameters has a value that isn't valid.

LimitExceededException

A limit is exceeded.

ListAnalysesError

ListDashboardVersionsError

ListDashboardsError

ListDataSetsError

ListDataSourcesError

ListFolderMembersError

ListFoldersError

ListGroupMembershipsError

ListGroupsError

ListIAMPolicyAssignmentsError

ListIAMPolicyAssignmentsForUserError

ListIngestionsError

ListNamespacesError

ListTagsForResourceError

ListTemplateAliasesError

ListTemplateVersionsError

ListTemplatesError

ListThemeAliasesError

ListThemeVersionsError

ListThemesError

ListUserGroupsError

ListUsersError

PreconditionNotMetException

One or more preconditions aren't met.

QuickSightUserNotFoundException

The user with the provided name isn't found. This error can happen in any operation

RegisterUserError

ResourceExistsException

The resource specified already exists.

ResourceNotFoundException

One or more resources can't be found.

ResourceUnavailableException

This resource is currently unavailable.

RestoreAnalysisError

SearchAnalysesError

SearchDashboardsError

SearchFoldersError

SessionLifetimeInMinutesInvalidException

The number of minutes specified for the lifetime of a session isn't valid. The session

TagResourceError

ThrottlingException

Access is throttled.

UnsupportedPricingPlanException

This error indicates that you are calling an embedding operation in Amazon QuickSight

UnsupportedUserEditionException

This error indicates that you are calling an operation on an Amazon QuickSight

UntagResourceError

UpdateAccountCustomizationError

UpdateAccountSettingsError

UpdateAnalysisError

UpdateAnalysisPermissionsError

UpdateDashboardError

UpdateDashboardPermissionsError

UpdateDashboardPublishedVersionError

UpdateDataSetError

UpdateDataSetPermissionsError

UpdateDataSourceError

UpdateDataSourcePermissionsError

UpdateFolderError

UpdateFolderPermissionsError

UpdateGroupError

UpdateIAMPolicyAssignmentError

UpdateTemplateAliasError

UpdateTemplateError

UpdateTemplatePermissionsError

UpdateThemeAliasError

UpdateThemeError

UpdateThemePermissionsError

UpdateUserError

Enums

CancelIngestionErrorKind

CreateAccountCustomizationErrorKind

CreateAnalysisErrorKind

CreateDashboardErrorKind

CreateDataSetErrorKind

CreateDataSourceErrorKind

CreateFolderErrorKind

CreateFolderMembershipErrorKind

CreateGroupErrorKind

CreateGroupMembershipErrorKind

CreateIAMPolicyAssignmentErrorKind

CreateIngestionErrorKind

CreateNamespaceErrorKind

CreateTemplateAliasErrorKind

CreateTemplateErrorKind

CreateThemeAliasErrorKind

CreateThemeErrorKind

DeleteAccountCustomizationErrorKind

DeleteAnalysisErrorKind

DeleteDashboardErrorKind

DeleteDataSetErrorKind

DeleteDataSourceErrorKind

DeleteFolderErrorKind

DeleteFolderMembershipErrorKind

DeleteGroupErrorKind

DeleteGroupMembershipErrorKind

DeleteIAMPolicyAssignmentErrorKind

DeleteNamespaceErrorKind

DeleteTemplateAliasErrorKind

DeleteTemplateErrorKind

DeleteThemeAliasErrorKind

DeleteThemeErrorKind

DeleteUserByPrincipalIdErrorKind

DeleteUserErrorKind

DescribeAccountCustomizationErrorKind

DescribeAccountSettingsErrorKind

DescribeAnalysisErrorKind

DescribeAnalysisPermissionsErrorKind

DescribeDashboardErrorKind

DescribeDashboardPermissionsErrorKind

DescribeDataSetErrorKind

DescribeDataSetPermissionsErrorKind

DescribeDataSourceErrorKind

DescribeDataSourcePermissionsErrorKind

DescribeFolderErrorKind

DescribeFolderPermissionsErrorKind

DescribeFolderResolvedPermissionsErrorKind

DescribeGroupErrorKind

DescribeIAMPolicyAssignmentErrorKind

DescribeIngestionErrorKind

DescribeNamespaceErrorKind

DescribeTemplateAliasErrorKind

DescribeTemplateErrorKind

DescribeTemplatePermissionsErrorKind

DescribeThemeAliasErrorKind

DescribeThemeErrorKind

DescribeThemePermissionsErrorKind

DescribeUserErrorKind

GenerateEmbedUrlForAnonymousUserErrorKind

GenerateEmbedUrlForRegisteredUserErrorKind

GetDashboardEmbedUrlErrorKind

GetSessionEmbedUrlErrorKind

ListAnalysesErrorKind

ListDashboardVersionsErrorKind

ListDashboardsErrorKind

ListDataSetsErrorKind

ListDataSourcesErrorKind

ListFolderMembersErrorKind

ListFoldersErrorKind

ListGroupMembershipsErrorKind

ListGroupsErrorKind

ListIAMPolicyAssignmentsErrorKind

ListIAMPolicyAssignmentsForUserErrorKind

ListIngestionsErrorKind

ListNamespacesErrorKind

ListTagsForResourceErrorKind

ListTemplateAliasesErrorKind

ListTemplateVersionsErrorKind

ListTemplatesErrorKind

ListThemeAliasesErrorKind

ListThemeVersionsErrorKind

ListThemesErrorKind

ListUserGroupsErrorKind

ListUsersErrorKind

RegisterUserErrorKind

RestoreAnalysisErrorKind

SearchAnalysesErrorKind

SearchDashboardsErrorKind

SearchFoldersErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAccountCustomizationErrorKind

UpdateAccountSettingsErrorKind

UpdateAnalysisErrorKind

UpdateAnalysisPermissionsErrorKind

UpdateDashboardErrorKind

UpdateDashboardPermissionsErrorKind

UpdateDashboardPublishedVersionErrorKind

UpdateDataSetErrorKind

UpdateDataSetPermissionsErrorKind

UpdateDataSourceErrorKind

UpdateDataSourcePermissionsErrorKind

UpdateFolderErrorKind

UpdateFolderPermissionsErrorKind

UpdateGroupErrorKind

UpdateIAMPolicyAssignmentErrorKind

UpdateTemplateAliasErrorKind

UpdateTemplateErrorKind

UpdateTemplatePermissionsErrorKind

UpdateThemeAliasErrorKind

UpdateThemeErrorKind

UpdateThemePermissionsErrorKind

UpdateUserErrorKind