Struct aws_sdk_quicksight::input::RegisterUserInput

#[non_exhaustive]
pub struct RegisterUserInput {
    pub identity_type: Option<IdentityType>,
    pub email: Option<String>,
    pub user_role: Option<UserRole>,
    pub iam_arn: Option<String>,
    pub session_name: Option<String>,
    pub aws_account_id: Option<String>,
    pub namespace: Option<String>,
    pub user_name: Option<String>,
    pub custom_permissions_name: Option<String>,
    pub external_login_federation_provider_type: Option<String>,
    pub custom_federation_provider_url: Option<String>,
    pub external_login_id: Option<String>,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

identity_type: Option<IdentityType>

Amazon QuickSight supports several ways of managing the identity of users. This parameter accepts two values:

IAM: A user whose identity maps to an existing IAMuser or role.

QUICKSIGHT: A user whose identity is owned and managed internally by Amazon QuickSight.

email: Option<String>

The email address of the user that you want to register.

user_role: Option<UserRole>

The Amazon QuickSight role for the user. The user role can be one of the following:

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon QuickSight settings.

RESTRICTED_READER: This role isn't currently available for use.

RESTRICTED_AUTHOR: This role isn't currently available for use.

iam_arn: Option<String>

The ARN of the IAMuser or role that you are registering with Amazon QuickSight.

session_name: Option<String>

You need to use this parameter only when you register one or more users using an assumed IAMrole. You don't need to provide the session name for other scenarios, for example when you are registering an IAMuser or an Amazon QuickSight user. You can register multiple users using the same IAMrole if each user has a different session name. For more information on assuming IAMroles, see assume-role in the AWS CLI Reference.

aws_account_id: Option<String>

The ID for the Amazon Web Services account that the user is in. Currently, you use the ID for the Amazon Web Services account that contains your Amazon QuickSight account.

namespace: Option<String>

The namespace. Currently, you should set this to default.

user_name: Option<String>

The Amazon QuickSight user name that you want to create for the user you are registering.

custom_permissions_name: Option<String>

(Enterprise edition only) The name of the custom permissions profile that you want to assign to this user. Customized permissions allows you to control a user's access by restricting access the following operations:

Create and update data sources

Create and update datasets

Create and update email reports

Subscribe to email reports

To add custom permissions to an existing user, use UpdateUser instead.

A set of custom permissions includes any combination of these restrictions. Currently, you need to create the profile names for custom permission sets by using the Amazon QuickSight console. Then, you use the RegisterUser API operation to assign the named set of permissions to a Amazon QuickSight user.

Amazon QuickSight custom permissions are applied through IAMpolicies. Therefore, they override the permissions typically granted by assigning Amazon QuickSight users to one of the default security cohorts in Amazon QuickSight (admin, author, reader).

This feature is available only to Amazon QuickSight Enterprise edition subscriptions.

external_login_federation_provider_type: Option<String>

The type of supported external login provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. The type of supported external login provider can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com. When choosing the COGNITO provider type, don’t use the "CustomFederationProviderUrl" parameter which is only needed when the external provider is custom.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider. When choosing CUSTOM_OIDC type, use the CustomFederationProviderUrl parameter to provide the custom OIDC provider URL.

custom_federation_provider_url: Option<String>

The URL of the custom OpenID Connect (OIDC) provider that provides identity to let a user federate into Amazon QuickSight with an associated Identity and Access Management(IAM) role. This parameter should only be used when ExternalLoginFederationProviderType parameter is set to CUSTOM_OIDC.

external_login_id: Option<String>

The identity ID for a user in the external login provider.

Implementations

impl RegisterUserInput

pub fn make_operation(
&self,
_config: &Config
) -> Result<Operation<RegisterUser, AwsErrorRetryPolicy>, BuildError>

Consumes the builder and constructs an Operation<RegisterUser>

pub fn builder() -> Builder

Creates a new builder-style object to manufacture RegisterUserInput