Struct aws_sdk_quicksight::input::update_user_input::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for UpdateUserInput

Implementations

impl Builder

pub fn user_name(self, input: impl Into<String>) -> Self

The Amazon QuickSight user name that you want to update.

pub fn set_user_name(self, input: Option<String>) -> Self

pub fn aws_account_id(self, input: impl Into<String>) -> Self

The ID for the Amazon Web Services account; that the user is in. Currently, you use the ID for the Amazon Web Services account; that contains your Amazon QuickSight account.

pub fn set_aws_account_id(self, input: Option<String>) -> Self

pub fn namespace(self, input: impl Into<String>) -> Self

The namespace. Currently, you should set this to default.

pub fn set_namespace(self, input: Option<String>) -> Self

pub fn email(self, input: impl Into<String>) -> Self

The email address of the user that you want to update.

pub fn set_email(self, input: Option<String>) -> Self

pub fn role(self, input: UserRole) -> Self

The Amazon QuickSight role of the user. The role can be one of the following default security cohorts:

READER: A user who has read-only access to dashboards.

AUTHOR: A user who can create data sources, datasets, analyses, and dashboards.

ADMIN: A user who is an author, who can also manage Amazon QuickSight settings.

The name of the QuickSight role is invisible to the user except for the console screens dealing with permissions.

pub fn set_role(self, input: Option<UserRole>) -> Self

pub fn custom_permissions_name(self, input: impl Into<String>) -> Self

(Enterprise edition only) The name of the custom permissions profile that you want to assign to this user. Customized permissions allows you to control a user's access by restricting access the following operations:

Create and update data sources

Create and update datasets

Create and update email reports

Subscribe to email reports

A set of custom permissions includes any combination of these restrictions. Currently, you need to create the profile names for custom permission sets by using the QuickSight console. Then, you use the RegisterUser API operation to assign the named set of permissions to a QuickSight user.

QuickSight custom permissions are applied through IAM policies. Therefore, they override the permissions typically granted by assigning QuickSight users to one of the default security cohorts in QuickSight (admin, author, reader).

This feature is available only to QuickSight Enterprise edition subscriptions.

pub fn set_custom_permissions_name(self, input: Option<String>) -> Self

pub fn unapply_custom_permissions(self, input: bool) -> Self

A flag that you use to indicate that you want to remove all custom permissions from this user. Using this parameter resets the user to the state it was in before a custom permissions profile was applied. This parameter defaults to NULL and it doesn't accept any other value.

pub fn set_unapply_custom_permissions(self, input: Option<bool>) -> Self

pub fn external_login_federation_provider_type(
self,
input: impl Into<String>
) -> Self

The type of supported external login provider that provides identity to let a user federate into QuickSight with an associated AWS Identity and Access Management (IAM) role. The type of supported external login provider can be one of the following.

COGNITO: Amazon Cognito. The provider URL is cognito-identity.amazonaws.com. When choosing the COGNITO provider type, don’t use the "CustomFederationProviderUrl" parameter which is only needed when the external provider is custom.

CUSTOM_OIDC: Custom OpenID Connect (OIDC) provider. When choosing CUSTOM_OIDC type, use the CustomFederationProviderUrl parameter to provide the custom OIDC provider URL.

NONE: This clears all the previously saved external login information for a user. Use DescribeUser API to check the external login information.

pub fn set_external_login_federation_provider_type(
self,
input: Option<String>
) -> Self

pub fn custom_federation_provider_url(self, input: impl Into<String>) -> Self

The URL of the custom OpenID Connect (OIDC) provider that provides identity to let a user federate into QuickSight with an associated AWS Identity and Access Management (IAM) role. This parameter should only be used when ExternalLoginFederationProviderType parameter is set to CUSTOM_OIDC.

pub fn set_custom_federation_provider_url(self, input: Option<String>) -> Self

pub fn external_login_id(self, input: impl Into<String>) -> Self

The identity ID for a user in the external login provider.

pub fn set_external_login_id(self, input: Option<String>) -> Self

pub fn build(self) -> Result<UpdateUserInput, BuildError>

Consumes the builder and constructs a UpdateUserInput