Module aws_sdk_quicksight::model

Modules

account_customization

See AccountCustomization

account_settings

See AccountSettings

active_iam_policy_assignment

See ActiveIamPolicyAssignment

ad_hoc_filtering_option

See AdHocFilteringOption

amazon_elasticsearch_parameters

See AmazonElasticsearchParameters

amazon_open_search_parameters

See AmazonOpenSearchParameters

analysis

See Analysis

analysis_error

See AnalysisError

analysis_search_filter

See AnalysisSearchFilter

analysis_source_entity

See AnalysisSourceEntity

analysis_source_template

See AnalysisSourceTemplate

analysis_summary

See AnalysisSummary

anonymous_user_dashboard_embedding_configuration

See AnonymousUserDashboardEmbeddingConfiguration

anonymous_user_embedding_experience_configuration

See AnonymousUserEmbeddingExperienceConfiguration

athena_parameters

See AthenaParameters

aurora_parameters

See AuroraParameters

aurora_postgre_sql_parameters

See AuroraPostgreSqlParameters

aws_iot_analytics_parameters

See AwsIotAnalyticsParameters

border_style

See BorderStyle

calculated_column

See CalculatedColumn

cast_column_type_operation

See CastColumnTypeOperation

column_description

See ColumnDescription

column_group

See ColumnGroup

column_group_column_schema

See ColumnGroupColumnSchema

column_group_schema

See ColumnGroupSchema

column_level_permission_rule

See ColumnLevelPermissionRule

column_schema

See ColumnSchema

column_tag

See ColumnTag

create_columns_operation

See CreateColumnsOperation

credential_pair

See CredentialPair

custom_sql

See CustomSql

dashboard

See Dashboard

dashboard_error

See DashboardError

dashboard_publish_options

See DashboardPublishOptions

dashboard_search_filter

See DashboardSearchFilter

dashboard_source_entity

See DashboardSourceEntity

dashboard_source_template

See DashboardSourceTemplate

dashboard_summary

See DashboardSummary

dashboard_version

See DashboardVersion

dashboard_version_summary

See DashboardVersionSummary

data_color_palette

See DataColorPalette

data_set

See DataSet

data_set_configuration

See DataSetConfiguration

data_set_reference

See DataSetReference

data_set_schema

See DataSetSchema

data_set_summary

See DataSetSummary

data_set_usage_configuration

See DataSetUsageConfiguration

data_source

See DataSource

data_source_credentials

See DataSourceCredentials

data_source_error_info

See DataSourceErrorInfo

date_time_parameter

See DateTimeParameter

decimal_parameter

See DecimalParameter

error_info

See ErrorInfo

export_to_csv_option

See ExportToCsvOption

field_folder

See FieldFolder

filter_operation

See FilterOperation

folder

See Folder

folder_member

See FolderMember

folder_search_filter

See FolderSearchFilter

folder_summary

See FolderSummary

geo_spatial_column_group

See GeoSpatialColumnGroup

group

See Group

group_member

See GroupMember

gutter_style

See GutterStyle

iam_policy_assignment

See IamPolicyAssignment

iam_policy_assignment_summary

See IamPolicyAssignmentSummary

ingestion

See Ingestion

input_column

See InputColumn

integer_parameter

See IntegerParameter

jira_parameters

See JiraParameters

join_instruction

See JoinInstruction

join_key_properties

See JoinKeyProperties

logical_table

See LogicalTable

logical_table_source

See LogicalTableSource

manifest_file_location

See ManifestFileLocation

margin_style

See MarginStyle

maria_db_parameters

See MariaDbParameters

member_id_arn_pair

See MemberIdArnPair

my_sql_parameters

See MySqlParameters

namespace_error

See NamespaceError

namespace_info_v2

See NamespaceInfoV2

oracle_parameters

See OracleParameters

output_column

See OutputColumn

parameters

See Parameters

postgre_sql_parameters

See PostgreSqlParameters

presto_parameters

See PrestoParameters

project_operation

See ProjectOperation

queue_info

See QueueInfo

rds_parameters

See RdsParameters

redshift_parameters

See RedshiftParameters

registered_user_dashboard_embedding_configuration

See RegisteredUserDashboardEmbeddingConfiguration

registered_user_embedding_experience_configuration

See RegisteredUserEmbeddingExperienceConfiguration

registered_user_quick_sight_console_embedding_configuration

See RegisteredUserQuickSightConsoleEmbeddingConfiguration

relational_table

See RelationalTable

rename_column_operation

See RenameColumnOperation

resource_permission

See ResourcePermission

row_info

See RowInfo

row_level_permission_data_set

See RowLevelPermissionDataSet

row_level_permission_tag_configuration

See RowLevelPermissionTagConfiguration

row_level_permission_tag_rule

See RowLevelPermissionTagRule

s3_parameters

See S3Parameters

s3_source

See S3Source

service_now_parameters

See ServiceNowParameters

session_tag

See SessionTag

sheet

See Sheet

sheet_controls_option

See SheetControlsOption

sheet_style

See SheetStyle

snowflake_parameters

See SnowflakeParameters

spark_parameters

See SparkParameters

sql_server_parameters

See SqlServerParameters

ssl_properties

See SslProperties

string_parameter

See StringParameter

tag

See Tag

tag_column_operation

See TagColumnOperation

template

See Template

template_alias

See TemplateAlias

template_error

See TemplateError

template_source_analysis

See TemplateSourceAnalysis

template_source_entity

See TemplateSourceEntity

template_source_template

See TemplateSourceTemplate

template_summary

See TemplateSummary

template_version

See TemplateVersion

template_version_summary

See TemplateVersionSummary

teradata_parameters

See TeradataParameters

theme

See Theme

theme_alias

See ThemeAlias

theme_configuration

See ThemeConfiguration

theme_error

See ThemeError

theme_summary

See ThemeSummary

theme_version

See ThemeVersion

theme_version_summary

See ThemeVersionSummary

tile_layout_style

See TileLayoutStyle

tile_style

See TileStyle

twitter_parameters

See TwitterParameters

ui_color_palette

See UiColorPalette

untag_column_operation

See UntagColumnOperation

upload_settings

See UploadSettings

user

See User

vpc_connection_properties

See VpcConnectionProperties

Structs

AccountCustomization

The Amazon QuickSight customizations associated with your Amazon Web Services account or a Amazon QuickSight namespace in a specific Amazon Web Services Region;.

AccountSettings

The Amazon QuickSight settings associated with your Amazon Web Services account.

ActiveIamPolicyAssignment

The active Identity and Access Management (IAM) policy assignment.

AdHocFilteringOption

Ad hoc (one-time) filtering option.

AmazonElasticsearchParameters

The parameters for Elasticsearch.

AmazonOpenSearchParameters

Analysis

Metadata structure for an analysis in Amazon QuickSight

AnalysisError

Analysis error.

AnalysisSearchFilter

A filter that you apply when searching for one or more analyses.

AnalysisSourceEntity

The source entity of an analysis.

AnalysisSourceTemplate

The source template of an analysis.

AnalysisSummary

The summary metadata that describes an analysis.

AnonymousUserDashboardEmbeddingConfiguration

Information about the dashboard that you want to embed.

AnonymousUserEmbeddingExperienceConfiguration

The type of experience you want to embed. For anonymous users, you can embed an Amazon QuickSight dashboard.

AthenaParameters

Parameters for Amazon Athena.

AuroraParameters

Parameters for Amazon Aurora.

AuroraPostgreSqlParameters

Parameters for Amazon Aurora PostgreSQL-Compatible Edition.

AwsIotAnalyticsParameters

The parameters for IoT Analytics.

BorderStyle

The display options for tile borders for visuals.

CalculatedColumn

A calculated column for a dataset.

CastColumnTypeOperation

A transform operation that casts a column to a different type.

ColumnDescription

Metadata that contains a description for a column.

ColumnGroup

Groupings of columns that work together in certain Amazon QuickSight features. This is

ColumnGroupColumnSchema

A structure describing the name, data type, and geographic role of the columns.

ColumnGroupSchema

The column group schema.

ColumnLevelPermissionRule

A rule defined to grant access on one or more restricted columns.

ColumnSchema

The column schema.

ColumnTag

A tag for a column in a TagColumnOperation structure. This is a

CreateColumnsOperation

A transform operation that creates calculated columns. Columns created in one such

CredentialPair

The combination of user name and password that are used as credentials.

CustomSql

A physical table type built from the results of the custom SQL query.

Dashboard

Dashboard.

DashboardError

Dashboard error.

DashboardPublishOptions

Dashboard publish options.

DashboardSearchFilter

A filter that you apply when searching for dashboards.

DashboardSourceEntity

Dashboard source entity.

DashboardSourceTemplate

Dashboard source template.

DashboardSummary

Dashboard summary.

DashboardVersion

Dashboard version.

DashboardVersionSummary

Dashboard version summary.

DataColorPalette

The theme colors that are used for data colors in charts. The colors description is a

DataSet

Dataset.

DataSetConfiguration

Dataset configuration.

DataSetReference

Dataset reference.

DataSetSchema

Dataset schema.

DataSetSummary

Dataset summary.

DataSetUsageConfiguration

The usage configuration to apply to child datasets that reference this dataset as a source.

DataSource

The structure of a data source.

DataSourceCredentials

Data source credentials. This is a variant type structure. For this structure to be

DataSourceErrorInfo

Error information for the data source creation or update.

DateTimeParameter

A date-time parameter.

DecimalParameter

A decimal parameter.

ErrorInfo

Error information for the SPICE ingestion of a dataset.

ExportToCsvOption

Export to .csv option.

FieldFolder

A FieldFolder element is a folder that contains fields and nested subfolders.

FilterOperation

A transform operation that filters rows based on a condition.

Folder

A folder.

FolderMember

An asset in a folder, such as a dashboard, analysis, or dataset.

FolderSearchFilter

Searches a folder by a filter.

FolderSummary

A summary of the folder.

GeoSpatialColumnGroup

Geospatial column group that denotes a hierarchy.

Group

A group in Amazon QuickSight consists of a set of users. You can

GroupMember

A member of an Amazon QuickSight group. Currently, group members must be users. Groups

GutterStyle

The display options for gutter spacing between tiles on a sheet.

IamPolicyAssignment

An Identity and Access Management (IAM) policy assignment.

IamPolicyAssignmentSummary

IAMpolicy assignment summary.

Ingestion

Information about the SPICE ingestion for a dataset.

InputColumn

Metadata for a column that is used as the input of a transform operation.

IntegerParameter

An integer parameter.

JiraParameters

The parameters for Jira.

JoinInstruction

The instructions associated with a join.

JoinKeyProperties

Properties associated with the columns participating in a join.

LogicalTable

A logical table is a unit that joins and that data

LogicalTableSource

Information about the source of a logical table. This is a variant type structure. For

ManifestFileLocation

Amazon S3 manifest file location.

MarginStyle

The display options for margins around the outside edge of sheets.

MariaDbParameters

The parameters for MariaDB.

MemberIdArnPair

An object that consists of the member Amazon Resource Name (ARN) and member ID.

MySqlParameters

The parameters for MySQL.

NamespaceError

Errors that occur during namespace creation.

NamespaceInfoV2

The error type.

OracleParameters

The parameters for Oracle.

OutputColumn

Output column.

Parameters

A list of Amazon QuickSight parameters and the list's override values.

PostgreSqlParameters

The parameters for PostgreSQL.

PrestoParameters

The parameters for Presto.

ProjectOperation

A transform operation that projects columns. Operations that come after a projection

QueueInfo

Information about a queued dataset SPICE ingestion.

RdsParameters

The parameters for Amazon RDS.

RedshiftParameters

The parameters for Amazon Redshift. The ClusterId field can be blank if

RegisteredUserDashboardEmbeddingConfiguration

Information about the dashboard you want to embed.

RegisteredUserEmbeddingExperienceConfiguration

The type of experience you want to embed. For registered users, you can embed an Amazon QuickSight dashboard or the Amazon QuickSight console.

RegisteredUserQuickSightConsoleEmbeddingConfiguration

Information about the Amazon QuickSight console that you want to embed.

RelationalTable

A physical table type for relational data sources.

RenameColumnOperation

A transform operation that renames a column.

ResourcePermission

Permission for the resource.

RowInfo

Information about rows for a data set SPICE ingestion.

RowLevelPermissionDataSet

Information about a dataset that contains permissions for row-level security (RLS).

RowLevelPermissionTagConfiguration

The configuration of tags on a dataset to set row-level security.

RowLevelPermissionTagRule

A set of rules associated with a tag.

S3Parameters

The parameters for S3.

S3Source

A physical table type for an S3 data source.

ServiceNowParameters

The parameters for ServiceNow.

SessionTag

The key-value pair used for the row-level security tags feature.

Sheet

A sheet, which is an object that contains a set of visuals that

SheetControlsOption

Sheet controls option.

SheetStyle

The theme display options for sheets.

SnowflakeParameters

The parameters for Snowflake.

SparkParameters

The parameters for Spark.

SqlServerParameters

The parameters for SQL Server.

SslProperties

Secure Socket Layer (SSL) properties that apply when Amazon QuickSight connects to your

StringParameter

A string parameter.

Tag

The key or keys of the key-value pairs for the resource tag or tags assigned to the

TagColumnOperation

A transform operation that tags a column with additional information.

Template

A template object. A template is an entity in Amazon QuickSight that

TemplateAlias

The template alias.

TemplateError

List of errors that occurred when the template version creation failed.

TemplateSourceAnalysis

The source analysis of the template.

TemplateSourceEntity

The source entity of the template.

TemplateSourceTemplate

The source template of the template.

TemplateSummary

The template summary.

TemplateVersion

A version of a template.

TemplateVersionSummary

The template version.

TeradataParameters

The parameters for Teradata.

Theme

Summary information about a theme.

ThemeAlias

An alias for a theme.

ThemeConfiguration

The theme configuration. This configuration contains all of the display properties for

ThemeError

Theme error.

ThemeSummary

The theme summary.

ThemeVersion

A version of a theme.

ThemeVersionSummary

The theme version.

TileLayoutStyle

The display options for the layout of tiles on a sheet.

TileStyle

Display options related to tiles on a sheet.

TwitterParameters

The parameters for Twitter.

UiColorPalette

The theme colors that apply to UI and to charts, excluding data colors. The colors

UntagColumnOperation

A transform operation that removes tags associated with a column.

UploadSettings

Information about the format for a source file or files.

User

A registered user of Amazon QuickSight.

VpcConnectionProperties

VPC connection properties.

Enums

AnalysisErrorType

AnalysisFilterAttribute

AssignmentStatus

ColumnDataType

ColumnTagName

DashboardBehavior

DashboardErrorType

DashboardFilterAttribute

DashboardUiState

DataSetImportMode

DataSourceErrorInfoType

NOTE: DataSourceErrorInfoType::Unknown has been renamed to ::UnknownValue.

DataSourceParameters

DataSourceType

Edition

EmbeddingIdentityType

ExceptionResourceType

FileFormat

FilterOperator

FolderFilterAttribute

FolderType

GeoSpatialCountryCode

GeoSpatialDataRole

IdentityStore

IdentityType

IngestionErrorType

IngestionRequestSource

IngestionRequestType

IngestionStatus

InputColumnDataType

JoinType

MemberType

NamespaceErrorType

NamespaceStatus

PhysicalTable

ResourceStatus

RowLevelPermissionFormatVersion

RowLevelPermissionPolicy

Status

TemplateErrorType

TextQualifier

ThemeErrorType

ThemeType

TransformOperation

UserRole