Enum aws_sdk_quicksight::model::DataSourceParameters

#[non_exhaustive]
pub enum DataSourceParameters {

    AmazonElasticsearchParameters(AmazonElasticsearchParameters),
    AmazonOpenSearchParameters(AmazonOpenSearchParameters),
    AthenaParameters(AthenaParameters),
    AuroraParameters(AuroraParameters),
    AuroraPostgreSqlParameters(AuroraPostgreSqlParameters),
    AwsIotAnalyticsParameters(AwsIotAnalyticsParameters),
    JiraParameters(JiraParameters),
    MariaDbParameters(MariaDbParameters),
    MySqlParameters(MySqlParameters),
    OracleParameters(OracleParameters),
    PostgreSqlParameters(PostgreSqlParameters),
    PrestoParameters(PrestoParameters),
    RdsParameters(RdsParameters),
    RedshiftParameters(RedshiftParameters),
    S3Parameters(S3Parameters),
    ServiceNowParameters(ServiceNowParameters),
    SnowflakeParameters(SnowflakeParameters),
    SparkParameters(SparkParameters),
    SqlServerParameters(SqlServerParameters),
    TeradataParameters(TeradataParameters),
    TwitterParameters(TwitterParameters),
    Unknown,

}

The parameters that Amazon QuickSight uses to connect to your underlying data source. This is a variant type structure. For this structure to be valid, only one of the attributes can be non-null.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AmazonElasticsearchParameters(AmazonElasticsearchParameters)

The parameters for OpenSearch.

Tuple Fields of AmazonElasticsearchParameters

0: AmazonElasticsearchParameters

AmazonOpenSearchParameters(AmazonOpenSearchParameters)

Tuple Fields of AmazonOpenSearchParameters

0: AmazonOpenSearchParameters

AthenaParameters(AthenaParameters)

The parameters for Amazon Athena.

Tuple Fields of AthenaParameters

0: AthenaParameters

AuroraParameters(AuroraParameters)

The parameters for Amazon Aurora MySQL.

Tuple Fields of AuroraParameters

0: AuroraParameters

AuroraPostgreSqlParameters(AuroraPostgreSqlParameters)

The parameters for Amazon Aurora.

Tuple Fields of AuroraPostgreSqlParameters

0: AuroraPostgreSqlParameters

AwsIotAnalyticsParameters(AwsIotAnalyticsParameters)

The parameters for IoT Analytics.

Tuple Fields of AwsIotAnalyticsParameters

0: AwsIotAnalyticsParameters

JiraParameters(JiraParameters)

The parameters for Jira.

Tuple Fields of JiraParameters

0: JiraParameters

MariaDbParameters(MariaDbParameters)

The parameters for MariaDB.

Tuple Fields of MariaDbParameters

0: MariaDbParameters

MySqlParameters(MySqlParameters)

The parameters for MySQL.

Tuple Fields of MySqlParameters

0: MySqlParameters

OracleParameters(OracleParameters)

The parameters for Oracle.

Tuple Fields of OracleParameters

0: OracleParameters

PostgreSqlParameters(PostgreSqlParameters)

The parameters for PostgreSQL.

Tuple Fields of PostgreSqlParameters

0: PostgreSqlParameters

PrestoParameters(PrestoParameters)

The parameters for Presto.

Tuple Fields of PrestoParameters

0: PrestoParameters

RdsParameters(RdsParameters)

The parameters for Amazon RDS.

Tuple Fields of RdsParameters

0: RdsParameters

RedshiftParameters(RedshiftParameters)

The parameters for Amazon Redshift.

Tuple Fields of RedshiftParameters

0: RedshiftParameters

S3Parameters(S3Parameters)

The parameters for S3.

Tuple Fields of S3Parameters

0: S3Parameters

ServiceNowParameters(ServiceNowParameters)

The parameters for ServiceNow.

Tuple Fields of ServiceNowParameters

0: ServiceNowParameters

SnowflakeParameters(SnowflakeParameters)

The parameters for Snowflake.

Tuple Fields of SnowflakeParameters

0: SnowflakeParameters

SparkParameters(SparkParameters)

The parameters for Spark.

Tuple Fields of SparkParameters

0: SparkParameters

SqlServerParameters(SqlServerParameters)

The parameters for SQL Server.

Tuple Fields of SqlServerParameters

0: SqlServerParameters

TeradataParameters(TeradataParameters)

The parameters for Teradata.

Tuple Fields of TeradataParameters

0: TeradataParameters

TwitterParameters(TwitterParameters)

The parameters for Twitter.

Tuple Fields of TwitterParameters

0: TwitterParameters

Unknown

The Unknown variant represents cases where new union variant was received. Consider upgrading the SDK to the latest available version. An unknown enum variant

Note: If you encounter this error, consider upgrading your SDK to the latest version. The Unknown variant represents cases where the server sent a value that wasn’t recognized by the client. This can happen when the server adds new functionality, but the client has not been updated. To investigate this, consider turning on debug logging to print the raw HTTP response.

Non-exhaustive enum variants could have additional fields added in future. Therefore, non-exhaustive enum variants cannot be constructed in external crates and cannot be matched against.

Implementations

impl DataSourceParameters

pub fn as_amazon_elasticsearch_parameters(
&self
) -> Result<&AmazonElasticsearchParameters, &Self>

Tries to convert the enum instance into AmazonElasticsearchParameters, extracting the inner AmazonElasticsearchParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_amazon_elasticsearch_parameters(&self) -> bool

Returns true if this is a AmazonElasticsearchParameters.

pub fn as_amazon_open_search_parameters(
&self
) -> Result<&AmazonOpenSearchParameters, &Self>

Tries to convert the enum instance into AmazonOpenSearchParameters, extracting the inner AmazonOpenSearchParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_amazon_open_search_parameters(&self) -> bool

Returns true if this is a AmazonOpenSearchParameters.

pub fn as_athena_parameters(&self) -> Result<&AthenaParameters, &Self>

Tries to convert the enum instance into AthenaParameters, extracting the inner AthenaParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_athena_parameters(&self) -> bool

Returns true if this is a AthenaParameters.

pub fn as_aurora_parameters(&self) -> Result<&AuroraParameters, &Self>

Tries to convert the enum instance into AuroraParameters, extracting the inner AuroraParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_aurora_parameters(&self) -> bool

Returns true if this is a AuroraParameters.

pub fn as_aurora_postgre_sql_parameters(
&self
) -> Result<&AuroraPostgreSqlParameters, &Self>

Tries to convert the enum instance into AuroraPostgreSqlParameters, extracting the inner AuroraPostgreSqlParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_aurora_postgre_sql_parameters(&self) -> bool

Returns true if this is a AuroraPostgreSqlParameters.

pub fn as_aws_iot_analytics_parameters(
&self
) -> Result<&AwsIotAnalyticsParameters, &Self>

Tries to convert the enum instance into AwsIotAnalyticsParameters, extracting the inner AwsIotAnalyticsParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_aws_iot_analytics_parameters(&self) -> bool

Returns true if this is a AwsIotAnalyticsParameters.

pub fn as_jira_parameters(&self) -> Result<&JiraParameters, &Self>

Tries to convert the enum instance into JiraParameters, extracting the inner JiraParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_jira_parameters(&self) -> bool

Returns true if this is a JiraParameters.

pub fn as_maria_db_parameters(&self) -> Result<&MariaDbParameters, &Self>

Tries to convert the enum instance into MariaDbParameters, extracting the inner MariaDbParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_maria_db_parameters(&self) -> bool

Returns true if this is a MariaDbParameters.

pub fn as_my_sql_parameters(&self) -> Result<&MySqlParameters, &Self>

Tries to convert the enum instance into MySqlParameters, extracting the inner MySqlParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_my_sql_parameters(&self) -> bool

Returns true if this is a MySqlParameters.

pub fn as_oracle_parameters(&self) -> Result<&OracleParameters, &Self>

Tries to convert the enum instance into OracleParameters, extracting the inner OracleParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_oracle_parameters(&self) -> bool

Returns true if this is a OracleParameters.

pub fn as_postgre_sql_parameters(&self) -> Result<&PostgreSqlParameters, &Self>

Tries to convert the enum instance into PostgreSqlParameters, extracting the inner PostgreSqlParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_postgre_sql_parameters(&self) -> bool

Returns true if this is a PostgreSqlParameters.

pub fn as_presto_parameters(&self) -> Result<&PrestoParameters, &Self>

Tries to convert the enum instance into PrestoParameters, extracting the inner PrestoParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_presto_parameters(&self) -> bool

Returns true if this is a PrestoParameters.

pub fn as_rds_parameters(&self) -> Result<&RdsParameters, &Self>

Tries to convert the enum instance into RdsParameters, extracting the inner RdsParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_rds_parameters(&self) -> bool

Returns true if this is a RdsParameters.

pub fn as_redshift_parameters(&self) -> Result<&RedshiftParameters, &Self>

Tries to convert the enum instance into RedshiftParameters, extracting the inner RedshiftParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_redshift_parameters(&self) -> bool

Returns true if this is a RedshiftParameters.

pub fn as_s3_parameters(&self) -> Result<&S3Parameters, &Self>

Tries to convert the enum instance into S3Parameters, extracting the inner S3Parameters. Returns Err(&Self) if it can’t be converted.

pub fn is_s3_parameters(&self) -> bool

Returns true if this is a S3Parameters.

pub fn as_service_now_parameters(&self) -> Result<&ServiceNowParameters, &Self>

Tries to convert the enum instance into ServiceNowParameters, extracting the inner ServiceNowParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_service_now_parameters(&self) -> bool

Returns true if this is a ServiceNowParameters.

pub fn as_snowflake_parameters(&self) -> Result<&SnowflakeParameters, &Self>

Tries to convert the enum instance into SnowflakeParameters, extracting the inner SnowflakeParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_snowflake_parameters(&self) -> bool

Returns true if this is a SnowflakeParameters.

pub fn as_spark_parameters(&self) -> Result<&SparkParameters, &Self>

Tries to convert the enum instance into SparkParameters, extracting the inner SparkParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_spark_parameters(&self) -> bool

Returns true if this is a SparkParameters.

pub fn as_sql_server_parameters(&self) -> Result<&SqlServerParameters, &Self>

Tries to convert the enum instance into SqlServerParameters, extracting the inner SqlServerParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_sql_server_parameters(&self) -> bool

Returns true if this is a SqlServerParameters.

pub fn as_teradata_parameters(&self) -> Result<&TeradataParameters, &Self>

Tries to convert the enum instance into TeradataParameters, extracting the inner TeradataParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_teradata_parameters(&self) -> bool

Returns true if this is a TeradataParameters.

pub fn as_twitter_parameters(&self) -> Result<&TwitterParameters, &Self>

Tries to convert the enum instance into TwitterParameters, extracting the inner TwitterParameters. Returns Err(&Self) if it can’t be converted.

pub fn is_twitter_parameters(&self) -> bool

Returns true if this is a TwitterParameters.

pub fn is_unknown(&self) -> bool

Returns true if the enum instance is the Unknown variant.