Module aws_sdk_quicksight::operation

Structs

CancelIngestion

Cancels an ongoing ingestion of data into SPICE.

CreateAccountCustomization

Creates Amazon QuickSight customizations the current Amazon Web Services Region;. Currently, you can

CreateAnalysis

Creates an analysis in Amazon QuickSight.

CreateDashboard

Creates a dashboard from a template. To first create a template, see the

CreateDataSet

Creates a dataset.

CreateDataSource

Creates a data source.

CreateFolder

Creates an empty shared folder.

CreateFolderMembership

Adds an asset, such as a dashboard, analysis, or dataset into a folder.

CreateGroup

Creates an Amazon QuickSight group.

CreateGroupMembership

Adds an Amazon QuickSight user to an Amazon QuickSight group.

CreateIAMPolicyAssignment

Creates an assignment with one specified IAMpolicy, identified by its Amazon Resource Name

CreateIngestion

Creates and starts a new SPICE ingestion on a dataset

CreateNamespace

(Enterprise edition only) Creates a new namespace for you to use with Amazon QuickSight.

CreateTemplate

Creates a template from an existing Amazon QuickSight analysis or template. You can use the resulting

CreateTemplateAlias

Creates a template alias for a template.

CreateTheme

Creates a theme.

CreateThemeAlias

Creates a theme alias for a theme.

DeleteAccountCustomization

Deletes all Amazon QuickSight customizations in this Amazon Web Services Region; for the specified

DeleteAnalysis

Deletes an analysis from Amazon QuickSight. You can optionally include a recovery window during

DeleteDashboard

Deletes a dashboard.

DeleteDataSet

Deletes a dataset.

DeleteDataSource

Deletes the data source permanently. This operation breaks

DeleteFolder

Deletes an empty folder.

DeleteFolderMembership

Removes an asset, such as a dashboard, analysis, or dataset, from a folder.

DeleteGroup

Removes a user group from Amazon QuickSight.

DeleteGroupMembership

Removes a user from a group so that the user is no longer a member of the group.

DeleteIAMPolicyAssignment

Deletes an existing IAMpolicy assignment.

DeleteNamespace

Deletes a namespace and the users and groups that are associated with the namespace.

DeleteTemplate

Deletes a template.

DeleteTemplateAlias

Deletes the item that the specified template alias points to. If you provide a specific

DeleteTheme

Deletes a theme.

DeleteThemeAlias

Deletes the version of the theme that the specified theme alias points to.

DeleteUser

Deletes the Amazon QuickSight user that is associated with the identity of the

DeleteUserByPrincipalId

Deletes a user identified by its principal ID.

DescribeAccountCustomization

Describes the customizations associated with the provided Amazon Web Services account and Amazon

DescribeAccountSettings

Describes the settings that were used when your Amazon QuickSight subscription was first

DescribeAnalysis

Provides a summary of the metadata for an analysis.

DescribeAnalysisPermissions

Provides the read and write permissions for an analysis.

DescribeDashboard

Provides a summary for a dashboard.

DescribeDashboardPermissions

Describes read and write permissions for a dashboard.

DescribeDataSet

Describes a dataset.

DescribeDataSetPermissions

Describes the permissions on a dataset.

DescribeDataSource

Describes a data source.

DescribeDataSourcePermissions

Describes the resource permissions for a data source.

DescribeFolder

Describes a folder.

DescribeFolderPermissions

Describes permissions for a folder.

DescribeFolderResolvedPermissions

Describes the folder resolved permissions. Permissions consists of both folder direct permissions and the inherited permissions from the ancestor folders.

DescribeGroup

Returns an Amazon QuickSight group's description and Amazon Resource Name (ARN).

DescribeIAMPolicyAssignment

Describes an existing IAMpolicy assignment, as specified by the assignment name.

DescribeIngestion

Describes a SPICE ingestion.

DescribeNamespace

Describes the current namespace.

DescribeTemplate

Describes a template's metadata.

DescribeTemplateAlias

Describes the template alias for a template.

DescribeTemplatePermissions

Describes read and write permissions on a template.

DescribeTheme

Describes a theme.

DescribeThemeAlias

Describes the alias for a theme.

DescribeThemePermissions

Describes the read and write permissions for a theme.

DescribeUser

Returns information about a user, given the user name.

GenerateEmbedUrlForAnonymousUser

Generates an embed URL that you can use to embed an Amazon QuickSight dashboard in your website, without having to register any reader users. Before you use this action, make sure that you have configured the dashboards and permissions.

GenerateEmbedUrlForRegisteredUser

Generates an embed URL that you can use to embed an Amazon QuickSight experience in your website. This action can be used for any type of user registered in an Amazon QuickSight account. Before you use this action, make sure that you have configured the relevant Amazon QuickSight resource and permissions.

GetDashboardEmbedUrl

Generates a session URL and authorization code that you can use to embed an Amazon

GetSessionEmbedUrl

Generates a session URL and authorization code that you can use to embed the Amazon

ListAnalyses

Lists Amazon QuickSight analyses that exist in the specified Amazon Web Services account.

ListDashboardVersions

Lists all the versions of the dashboards in the Amazon QuickSight subscription.

ListDashboards

Lists dashboards in an Amazon Web Services account.

ListDataSets

Lists all of the datasets belonging to the current Amazon Web Services account in an Amazon Web Services Region;.

ListDataSources

Lists data sources in current Amazon Web Services Region; that belong to this Amazon Web Services account.

ListFolderMembers

List all assets (DASHBOARD, ANALYSIS, and DATASET) in a folder.

ListFolders

Lists all folders in an account.

ListGroupMemberships

Lists member users in a group.

ListGroups

Lists all user groups in Amazon QuickSight.

ListIAMPolicyAssignments

Lists IAMpolicy assignments in the current Amazon QuickSight account.

ListIAMPolicyAssignmentsForUser

Lists all the IAMpolicy assignments, including the Amazon Resource Names (ARNs) for the IAM

ListIngestions

Lists the history of SPICE ingestions for a dataset.

ListNamespaces

Lists the namespaces for the specified Amazon Web Services account.

ListTagsForResource

Lists the tags assigned to a resource.

ListTemplateAliases

Lists all the aliases of a template.

ListTemplateVersions

Lists all the versions of the templates in the current Amazon QuickSight account.

ListTemplates

Lists all the templates in the current Amazon QuickSight account.

ListThemeAliases

Lists all the aliases of a theme.

ListThemeVersions

Lists all the versions of the themes in the current Amazon Web Services account.

ListThemes

Lists all the themes in the current Amazon Web Services account.

ListUserGroups

Lists the Amazon QuickSight groups that an Amazon QuickSight user is a member of.

ListUsers

Returns a list of all of the Amazon QuickSight users belonging to this account.

RegisterUser

Creates an Amazon QuickSight user, whose identity is associated with the AWS Identity

RestoreAnalysis

Restores an analysis.

SearchAnalyses

Searches for analyses that belong to the user specified in the filter.

SearchDashboards

Searches for dashboards that belong to a user.

SearchFolders

Searches the subfolders in a folder.

TagResource

Assigns one or more tags (key-value pairs) to the specified Amazon QuickSight resource.

UntagResource

Removes a tag or tags from a resource.

UpdateAccountCustomization

Updates Amazon QuickSight customizations the current Amazon Web Services Region;. Currently, the only

UpdateAccountSettings

Updates the Amazon QuickSight settings in your Amazon Web Services account.

UpdateAnalysis

Updates an analysis in Amazon QuickSight

UpdateAnalysisPermissions

Updates the read and write permissions for an analysis.

UpdateDashboard

Updates a dashboard in an Amazon Web Services account.

UpdateDashboardPermissions

Updates read and write permissions on a dashboard.

UpdateDashboardPublishedVersion

Updates the published version of a dashboard.

UpdateDataSet

Updates a dataset.

UpdateDataSetPermissions

Updates the permissions on a dataset.

UpdateDataSource

Updates a data source.

UpdateDataSourcePermissions

Updates the permissions to a data source.

UpdateFolder

Updates the name of a folder.

UpdateFolderPermissions

Updates permissions of a folder.

UpdateGroup

Changes a group description.

UpdateIAMPolicyAssignment

Updates an existing IAMpolicy assignment. This operation updates only the optional

UpdateTemplate

Updates a template from an existing Amazon QuickSight analysis or another template.

UpdateTemplateAlias

Updates the template alias of a template.

UpdateTemplatePermissions

Updates the resource permissions for a template.

UpdateTheme

Updates a theme.

UpdateThemeAlias

Updates an alias of a theme.

UpdateThemePermissions

Updates the resource permissions for a theme. Permissions apply to the action to grant or

UpdateUser

Updates an Amazon QuickSight user.